+++ mongodb/Changelog.md
@@ -1,4 +1,5 @@
 0.1.0 (Unreleased)
 ==================
+- Agent metrics.
 - Basic info and sharding data for MongoDB 3.2+
 - OpenTracing integration.